+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 - `lib/message`: `Report()` can now create a Report instance from Event instances (#1225).
 - `lib/bot`: The first word in the log line `Processed ... messages since last logging.` is now adaptible and set to `Forwarded` in the existing filtering bots (#1237).
 - `lib/bot`: Kills oneself again after proper shutdown if the bot is XMPP collector or output (#970). Previously these two bots needed two stop commands to get actually stopped.
+- `lib/utils`: log: set the name of the `py.warnings` logger to the bot name (#1184).
 
 ### Harmonization